+++ packages/khao-ui/vite.build.config.ts
@@ -18,7 +18,7 @@ export default defineConfig({
         chip: "src/components/misc/chip/Chip.svelte",
         "content-with-icon":
           "src/components/misc/contentWithIcon/ContentWithIcon.svelte",
-        "copy-tp-clipboard":
+        "copy-to-clipboard":
           "src/components/buttons/copyToClipboard/CopyToClipboard.svelte",
         icon: "src/icons/Icon.svelte",
         "icon-button": "src/components/buttons/iconButton/IconButton.svelte",